refactor(gameobject): forward args from instantiate
Arguments passed to the instantiate function are now passed on to the
initialize function of GameObjects. This solves the issue where objects
would need to define another initialize function that takes different
arguments, instead of passing it into the standard initialize function.
Arguments passed to instantiate must match (one of) the objects
initialize function(s).

Change bullet to use shared_ptrs instead of raw pointers to a unique_ptr
owned by the player, to ensure that the GunData object is available.