fix(terrain): wrong pixel removed on bullet collision
Fixes #120

Problem:
Due to how the pixel to remove was calculated it could sometimes remove
one that should not be able to be hit, for example when shooting
diagonally it would register behind the two pixels one would expect to
be hit.

Solution:
Check if the calculated position differs from the collision position on
more than one axis, and if so choose the one that differs the least.