feat(terrain): create random details for shape corners
Trying to see what the results are like if we first generate a
low-resolution "shape" map, which will be used for the overall layout
and shape of the terrain, and then generate details from there.
This detects if a part of an empty shape-cell is in a corner, and fills
the corresponding corner with random terrain.