feat(terrain): break up consecutive walls
Changed the function for making random "details" along consecutive walls
for the final pass to a generic function that can be used to break up
consecutive walls at any stage.
This new function is used at the end of generating the shape to have a
larger impact on how the overall terrain looks.

Also added public functions for setting all parameters of the
generation.