refactor: make TerrainCollider a LineCollider
Terrain collision now works as before, but the collider is not part of a
GameObject and is instead owned by the chunk.
Also add possibility to add tags to colliders. This is used in
TerrainCollider to give it the tag "Terrain", so that other objects can
know that they collided with terrain. The collider no longer has a
GameObject parent, so this is necessary.