fix: TerrainCollider fake GameObject memory issue (#151)
Fixes #148 

Problem:
It seems that the TerrainManager was being copied, which caused a
shallow copy of the TerrainColliders resulting in a double free. This
was only apparent when fakeGameObject was not null, because then (I
believe) the destructor was no longer trivial.

Solution:
Delete copy constructors and force move construction for
TerrainCollider, Chunk, and TerrainManager.